+++ apps/multiplatform/common/src/commonMain/kotlin/chat/simplex/common/views/usersettings/SettingsView.kt
@@ -32,6 +32,7 @@ import chat.simplex.common.views.helpers.*
 import chat.simplex.common.views.migration.MigrateFromDeviceView
 import chat.simplex.common.views.onboarding.SimpleXInfo
 import chat.simplex.common.views.onboarding.WhatsNewView
+import chat.simplex.common.views.usersettings.networkAndServers.NetworkAndServersView
 import chat.simplex.res.MR
 
 @Composable
@@ -154,6 +155,39 @@ fun HelpAndSupportView(
     SectionBottomSpacer()
   }
 }
+
+@Composable
+fun AdvancedSettingsView(
+  chatModel: ChatModel,
+  showModal: (@Composable (ChatModel) -> Unit) -> (() -> Unit),
+  showSettingsModal: (@Composable (ChatModel) -> Unit) -> (() -> Unit),
+  showCustomModal: (@Composable ModalData.(ChatModel, () -> Unit) -> Unit) -> (() -> Unit),
+  withAuth: (title: String, desc: String, block: () -> Unit) -> Unit,
+) {
+  val stopped = chatModel.chatRunning.value == false
+  val notificationsMode = remember { chatModel.controller.appPrefs.notificationsMode.state }
+  ColumnWithScrollBar {
+    AppBarTitle(stringResource(MR.strings.advanced_settings))
+
+    SectionView {
+      SettingsActionItem(painterResource(MR.images.ic_wifi_tethering), stringResource(MR.strings.network_and_servers), showCustomModal { _, close -> NetworkAndServersView(close) }, disabled = stopped)
+      if (appPlatform == AppPlatform.ANDROID) {
+        SettingsActionItem(painterResource(if (notificationsMode.value == NotificationsMode.OFF) MR.images.ic_bolt_off else MR.images.ic_bolt), stringResource(MR.strings.notifications), showSettingsModal { NotificationsSettingsView(it) }, disabled = stopped)
+      }
+      SettingsActionItem(painterResource(MR.images.ic_videocam), stringResource(MR.strings.settings_audio_video_calls), showSettingsModal { CallSettingsView(it, showModal) }, disabled = stopped)
+    }
+    SectionDividerSpaced()
+
+    AdvancedSettingsAppSection(showSettingsModal, withAuth)
+    SectionBottomSpacer()
+  }
+}
+
+@Composable
+expect fun AdvancedSettingsAppSection(
+  showSettingsModal: (@Composable (ChatModel) -> Unit) -> (() -> Unit),
+  withAuth: (title: String, desc: String, block: () -> Unit) -> Unit,
+)
 
 @Composable private fun DatabaseItem(encrypted: Boolean, saved: Boolean, openDatabaseView: () -> Unit, stopped: Boolean) {
   SectionItemView(openDatabaseView) {
